requirements.txt
pillow
pickleshare
numpy
tensorflow-cpu
Flask
urllib3
gunicorn
matplotlib
